+++ CHANGES.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.9.0 WIP
 - Group by levels is now default on True
--  
+- Add since season date on rating page (default last 2 seasons) 
 
 ## 0.8.0
 - FIX: When splinterlands is in maintenance mode skip update, causing application crashes